feat: Dialog component, alert function for easy usability (#129)
components now export an `alert` function, that takes almost identical
arguments as React Native's `Alert.alert`, with the possibility to add
icons to the buttons. As long as the app uses EDSProvider, the user
should be able to use our `alert` function. We also export the Dialog
component, which has almost identical API as `eds-core-react`. However,
it is not recommended to use that component unless `alert` does not
satisfy the user's needs.

New alert function vs React native alert:

![image](https://github.com/equinor/mad/assets/22524448/1b5be762-3cdb-41bc-a6f5-2d192c9b2024)

iPhone 8 vs iPad Mini:

![image](https://github.com/equinor/mad/assets/22524448/a75924ba-13b0-43d4-89e6-37ea416b9c83)

Please test this locally and see what you think. Also, I am constantly
getting this error both in main and in this branch. Don't know if this
is an issue only on my machine, or something I've accidentally
introduced

![image](https://github.com/equinor/mad/assets/22524448/03f7e18e-67f4-433e-972a-cdf9cd94dce9)

---------

Co-authored-by: Haakon Svane <36899138+HaakonSvane@users.noreply.github.com>